+++ FirebaseAI/Sources/Types/Public/SpeechConfig.swift
@@ -14,31 +14,51 @@
 
 import Foundation
 
-/// Configuration for controlling the voice of the model during conversation.
-@available(watchOS, unavailable)
+/// Speech configuration class for controlling the model's speech and audio generation behaviors.
+///
+/// This allows you to configure the voice properties (single-speaker OR multi-speaker setup) and
+/// language preferences when requesting the model to generate spoken responses.
 public struct SpeechConfig: Sendable {
-  let speechConfig: BidiSpeechConfig
+  let speechConfig: ProtoSpeechConfig
 
-  init(_ speechConfig: BidiSpeechConfig) {
+  init(_ speechConfig: ProtoSpeechConfig) {
     self.speechConfig = speechConfig
   }
 
-  /// Creates a new ``SpeechConfig`` value.
+  /// Creates a new ``SpeechConfig`` value for a single voice.
   ///
   /// - Parameters:
   ///   - voiceName: The name of the prebuilt voice to be used for the model's speech response.
   ///
   ///     To learn more about the available voices, see the docs on
   ///     [Voice options](https://ai.google.dev/gemini-api/docs/speech-generation#voices)\.
-  ///   - languageCode: ISO-639 language code to use when parsing text sent from the client, instead
+  ///   - languageCode: BCP-47 language code to use when parsing text sent from the client, instead
   ///     of audio. By default, the model will attempt to detect the input language automatically.
   ///
   ///     To learn which codes are supported, see the docs on
   ///     [Supported languages](https://ai.google.dev/gemini-api/docs/speech-generation#languages)\.
   public init(voiceName: String, languageCode: String? = nil) {
     self.init(
-      BidiSpeechConfig(
+      ProtoSpeechConfig(
         voiceConfig: .prebuiltVoiceConfig(.init(voiceName: voiceName)),
+        languageCode: languageCode
+      )
+    )
+  }
+
+  /// Creates a new ``SpeechConfig`` value for a multi-speaker setup.
+  ///
+  /// > Warning: Multi-speaker configurations are not currently supported by the Live API (e.g.,
+  /// > `LiveGenerationConfig`).
+  ///
+  /// - Parameters:
+  ///   - multiSpeakerVoiceConfig: The configuration detailing multiple speakers and their
+  ///     corresponding voices.
+  ///   - languageCode: BCP-47 language code to use when parsing text sent from the client.
+  public init(multiSpeakerVoiceConfig: MultiSpeakerVoiceConfig, languageCode: String? = nil) {
+    self.init(
+      ProtoSpeechConfig(
+        multiSpeakerVoiceConfig: multiSpeakerVoiceConfig.multiSpeakerVoiceConfig,
         languageCode: languageCode
       )
     )